Refactored IErrorReportingService to extract the info bar display logic into IInfoBarService
