+++ src/VisualStudio/Core/Def/Implementation/Workspace/VisualStudioErrorReportingServiceFactory.cs
@@ -1,15 +1,9 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System;
-using System.Collections.Generic;
 using System.Composition;
-using Microsoft.CodeAnalysis.Editor;
 using Microsoft.CodeAnalysis.Extensions;
 using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.Host.Mef;
-using Microsoft.CodeAnalysis.Shared.TestHooks;
-using Microsoft.VisualStudio.LanguageServices.Implementation.ProjectSystem;
-using Microsoft.VisualStudio.Shell;
 
 namespace Microsoft.VisualStudio.LanguageServices.Implementation
 {
@@ -19,12 +13,9 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation
         private readonly IErrorReportingService _singleton;
 
         [ImportingConstructor]
-        public VisualStudioErrorReportingServiceFactory(
-            SVsServiceProvider serviceProvider,
-            IForegroundNotificationService foregroundNotificationService,
-            [ImportMany] IEnumerable<Lazy<IAsynchronousOperationListener, FeatureMetadata>> asyncListeners)
+        public VisualStudioErrorReportingServiceFactory(IInfoBarService infoBarService)
         {
-            _singleton = new VisualStudioErrorReportingService(serviceProvider, foregroundNotificationService, new AggregateAsynchronousOperationListener(asyncListeners, FeatureAttribute.InfoBar));
+            _singleton = new VisualStudioErrorReportingService(infoBarService);
         }
 
         public IWorkspaceService CreateService(HostWorkspaceServices workspaceServices)
